Rename RemoteExecutor module to remoteexecutor
